requirements.txt
# mysql-connector-python==8.0.33
# bcrypt==4.0.1
# requests==2.31.0
# python-dotenv==1.0.0

mysql-connector-python==8.0.31
bcrypt==3.2.0
requests==2.28.1
python-dotenv==0.21.0
